+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Reduce landmark model memory usage.
 - Remove the experimental flag from `raster-allow-draping`.
 - Add support for emissive MRT in custom layers.
+- Improve style parsing performance.
 
 ### Bug fixes 🐞